fix short distance

+++ common/measureRange/measureRange.h
@@ -8,10 +8,10 @@
 
 class MeasureRange {
   private:
-    uint32_t range;
+    uint16_t range;
     uint8_t measureType;;
   public:
     uint64_t serialize();
     void deserialize(uint8_t data[8]);
-    void set(uint32_t range, uint8_t measureType);
+    void set(uint16_t range, uint8_t measureType);
 };
